requirements.txt
google-cloud-aiplatform
google-auth-oauthlib
google-cloud-secret-manager
google-cloud-firestore
flask
pydantic
presidio-analyzer
presidio-anonymizer
pytest
pytest-asyncio
google-api-python-client
googlemaps
cryptography
msgraph-core
azure-identity
msal
firebase-admin
pyjwt
flask-cors
stripe
